+++ unikmer/cmd/grep.go
@@ -67,7 +67,7 @@ Tips:
 		if opt.Verbose {
 			log.Info("checking input files ...")
 		}
-		files := getFileListFromArgsAndFile(cmd, args, true, "infile-list", true)
+		files := getFileListFromArgsAndFile(cmd, args, true, "infile-list", !opt.SkipFileCheck)
 		if opt.Verbose {
 			if len(files) == 1 && isStdin(files[0]) {
 				log.Info("no files given, reading from stdin")