+++ unikmer/cmd/inter.go
@@ -59,7 +59,7 @@ Tips:
 		if opt.Verbose {
 			log.Info("checking input files ...")
 		}
-		files := getFileListFromArgsAndFile(cmd, args, true, "infile-list", true)
+		files := getFileListFromArgsAndFile(cmd, args, true, "infile-list", !opt.SkipFileCheck)
 		if opt.Verbose {
 			if len(files) == 1 && isStdin(files[0]) {
 				log.Info("no files given, reading from stdin")
@@ -121,7 +121,7 @@ Tips:
 
 		// checking files
 		for _, file := range files {
-			if opt.NoCheckFile {
+			if opt.SkipFlagCheck {
 				break
 			}